Dont assume the only overflow we'll ever deal with is usize

+++ primes/src/errors/mod.rs
@@ -2,7 +2,7 @@ use std::{fmt, io};
 
 pub enum Error {
     InvalidAlgorithm,
-    MaxOverflow(usize),
+    MaxOverflow(u64),
     IO(String),
 }
